Handle inputs more clearly
The review requested to sanitize the user-provided input before passing
it on. As the UserAgent as well as the IP address are not really stored
or executed anywhere but merely used to match a regex agsint this is not
only not relevant but might actually falsify the matching as possible
"special values" that are the result of a santitzation might break the
matching.

This clairifes that in a comment but also adds a verification of the
IP-address to make sure it is a valid IP that belongs to a globally
accessible range. We do not use private IP-addresses.